Feat(Delete Contact Dialog Agreement)

+++ src/views/main/contacts/contact_list_items/index.tsx
@@ -6,20 +6,32 @@ import { List } from '@material-ui/core'
 import ContactFormDialog from '../contact_dialog_form'
 import ContactService from '../../../../services/contact/ContactService'
 import Constants from '../../../../constants/contact/ContactsConstants'
+import AgreementDialog from '../../../../components/agreement_dialog'
+import { useCurrentLanguage } from '../../../../context/provider/app_settings'
 
 const ContactItems = ({ items, setItems }: ContactItemsProps): JSX.Element => {
+
   const [cardOpen, setCardOpen] = useState(0)
   const [edit, setEdit] = useState(0)
   const [_delete, setDelete] = useState(0)
+  const [deleteDialogAgreement, setDeleteDialogAgreement] = useState(false)
+
+  const language = useCurrentLanguage()
 
   useEffect(() => {
-    if (_delete) {
+    if (_delete && deleteDialogAgreement) {
       ContactService.deleteContact(_delete)
-      setDelete(0)
+      handleCloseDeleteDialog()
     }
-  }, [_delete, setItems])
+  }, [_delete, deleteDialogAgreement])
+
+  const handleCloseDeleteDialog = () => {
+    setDeleteDialogAgreement(false)
+    setDelete(0)
+  }
 
   return (
+    <>
     <List className="contacts__list">
       {items.map((contact, index) => (
         <div key={index}>
@@ -46,6 +58,16 @@ const ContactItems = ({ items, setItems }: ContactItemsProps): JSX.Element => {
         </div>
       ))}
     </List>
+    <AgreementDialog 
+      open={Boolean(_delete)}
+      agreeOptionText={language.AGREEMENT_OPTION_TEXT}
+      disagreeOptionText={language.DISAGREEMENT_OPTION_TEXT}
+      description={language.DELETE_CONTACT_OPTION_TEXT}
+      question={language.DELETE_CONTACT_QUESTION}
+      onAgree={() => setDeleteDialogAgreement(true)}
+      onDisagree={handleCloseDeleteDialog}
+    />
+  </>
   )
 }
